Add files via upload
the dataset of English Vocabulary frequency. the dataset is scraped from the internet, so it contains unwanted words such as abbreviated forms of 'are' and 'have' ('re, 've) but without the apostrophe ('). so had to clean the file by dropping those data points. the out.csv is the resulted cleaned dataset preferably used.